Schedule tasks to APs in `aster-nix`
